requirements.txt
anthropic>=0.40.0
feedparser==6.0.11
trafilatura==1.12.2
google-cloud-texttospeech>=2.17.0
pydub>=0.25.1
PyYAML>=6.0.1
requests>=2.32.0
python-dateutil>=2.9.0
feedgen>=1.0.0
python-dotenv>=1.0.0
